Update times to 24 hour clock
I have a colorie spreadsheet that has times without am or pm specified.
Google Sheets interprets all the times as am. If I update the times to
24-hour times then Google Sheets knows what to do. This escript reads in
each line and determines if the time should be updated.

The first times in each day are assumed to be am (I always eat in the
morning).
Any time a time is lower than a previous time I assume it's afternoon.
Any line that doesn't start with a time is assumed to be part of a new
day and am/pm is reset to am.