Chemical Signalling
I was reading "Reinventing Man: the robot becomes reality" and got
inspired to model some chemical signaling. The idea is that molecules
are coming through the blood stream to a receptor site. The more
receptors that fill up the stronger the nerve signal gets. The receptor
only has a certain number of receptors so the strength gets maxed out
when the receptors all get full. The molecules only last for so long
before disappearing (for whatever reason) and the signal starts to
degrade. The rates of the molecules entering the blood stream and the
"average" age of the molecules can be controlled.

c("chem_sig.erl").
chem_sig:run().
producer ! {rate, 900}.