Print out all the chunks
tEXt chunks are printed out with the 0 byte replaced by a semi-colon and
without chunk info.

All other chunks are printed out with a maximum of 100 bytes of data as
well as the chunk info like data length, header type, whether or not
this is an "extra" or "ancillary" chunk, etc.