Dynamically add children to one-for-one sup
Somewhere along the line I got the impression that a simple-one-for-one
supervisor wouldn't restart it's children. I wanted to see if I could
use a one-for-one supervisor and add the children dynamically. You can.
However, I can't find anywhere in the documentation--and it doesn't make
sense-- that says a simple-one-for-one supervisor won't restart it's
chidlren.