Render path in second shell
I want to use a second shell with an alternate screen buffer to animate
the path finding. Normally I'd animate in JavaScript through a web
socket, but this seems like more fun.

Shell 1:
1) Start up a shell as 'port@<host>'
2) run `plot:plot()`.

Shell 2:
3) Start up a different shell
4) send {graph, W, H} to {port, 'port@<host>'} to draw the graph
5) send {X, Y} for each path step.

Shell 1 will draw the graph ... but I have to implement drawing in each
path step and the interior walls.